requirements.txt
Flask>=2.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0    # optional for local testing with .env
